review: address code review feedback for PR5 trace instrumentation
I1: Fix DisaggregationRecord doc comment to enumerate all three drop
paths that produce an orphaned record (empty prefill pool, empty decode
pool, AllocateTransferredKV failure) instead of only the third.

I2: Remove incorrect R1 rule citation from pd_events.go trace placement
comment. R1 covers silent data loss via counters, not trace record
placement. Replaced with a plain-language description of the placement
invariant.

I3: Add TestPDTrace_NeverDecider_WithPools_OnlyDisaggRecords — verifies
that with PDDecider="never" and pools configured, DisaggregationDecision
Event fires and records decisions with Disaggregate=false, and no
PrefillRouting, KVTransfer, or DecodeRouting records are emitted.

I4: Add TestPDTrace_DroppedAtDecodeKV_NoOrphanRecords — verifies that
when AllocateTransferredKV fails (decode KV OOM), no KVTransferRecord or
DecodeRoutingRecord is emitted for the dropped request, exercising the
"placement after AllocateTransferredKV" trace invariant end-to-end.

I5: Add three TraceSummary PD field unit tests to summary_test.go:
TestSummarize_PDFields_DisaggregationCounting (exact true/false counts),
TestSummarize_PDFields_KVTransferMean (exact mean computation),
TestSummarize_PDFields_EmptyPDRecords_ZeroValues (no PD records → zeros).

S1-S4: Minor comment improvements — UniqueTargets scope note, clarified
DisaggregatedCount/KVTransferCount field descriptions, MeanTransferDuration
zero-when-empty note, ParentRequestID/RequestID join asymmetry note on
PrefillRoutingRecord and DecodeRoutingRecord, float64 accumulation in
MeanTransferDuration to avoid int64 overflow.

+++ sim/trace/summary.go
@@ -8,13 +8,13 @@ type TraceSummary struct {
 	MeanRegret         float64
 	MaxRegret          float64
 	UniqueTargets      int
-	TargetDistribution map[string]int // instance ID → count of requests routed via standard routing only (not PD pool routing)
+	TargetDistribution map[string]int // instance ID → count of requests routed via standard routing only (not PD pool routing); use PrefillRoutings/DecodeRoutings for per-pool counts
 
 	// PD disaggregation summary (zero when disaggregation is not configured)
-	DisaggregationCount  int     // number of disaggregation decisions recorded
-	DisaggregatedCount   int     // number of requests routed to prefill pool (Disaggregate=true)
-	KVTransferCount      int     // number of completed KV transfers
-	MeanTransferDuration float64 // mean KV transfer duration in microseconds
+	DisaggregationCount  int     // number of disaggregation decisions recorded (true and false combined)
+	DisaggregatedCount   int     // number of requests for which disaggregation was decided (Disaggregate=true); prefill routing happens in a subsequent event
+	KVTransferCount      int     // number of KV transfers that completed with successful decode KV allocation
+	MeanTransferDuration float64 // mean KV transfer duration in microseconds; zero when KVTransferCount == 0
 }
 
 // Summarize computes aggregate statistics from a SimulationTrace.
@@ -48,6 +48,7 @@ func Summarize(st *SimulationTrace) *TraceSummary {
 		summary.MeanRegret = totalRegret / float64(len(st.Routings))
 	}
 
+	// UniqueTargets counts distinct instances in TargetDistribution (standard routing only, not PD pool routing).
 	summary.UniqueTargets = len(summary.TargetDistribution)
 
 	// PD disaggregation summary
@@ -60,11 +61,13 @@ func Summarize(st *SimulationTrace) *TraceSummary {
 
 	summary.KVTransferCount = len(st.KVTransfers)
 	if len(st.KVTransfers) > 0 {
-		totalDuration := int64(0)
+		// Accumulate in float64 to avoid int64 overflow for large simulations with many
+		// long-duration transfers (int64 max ~2.9×10^11 seconds; float64 is safe to ~10^15 µs).
+		totalDuration := 0.0
 		for _, kv := range st.KVTransfers {
-			totalDuration += kv.TransferDuration
+			totalDuration += float64(kv.TransferDuration)
 		}
-		summary.MeanTransferDuration = float64(totalDuration) / float64(len(st.KVTransfers))
+		summary.MeanTransferDuration = totalDuration / float64(len(st.KVTransfers))
 	}
 
 	return summary
